pyproject.toml
[tool.poetry]
name = "travel-messanger"
version = "0.1.0"
description = ""
authors = ["DmiKul, Neverless, 3str-webdev"]
license = "MIT"
readme = "README.md"
packages = [{include = "travel_messanger"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.7"
fastapi = "^0.95.0"
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^10.4.1"}
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
